+++ tests/test_sql_utils.py
@@ -1,7 +1,8 @@
 import unittest
 
 from streaming_jupyter_integrations.sql_utils import (inline_sql_in_cell,
-                                                      is_ddl, is_dml, is_dql)
+                                                      is_ddl, is_dml, is_dql,
+                                                      is_metadata_query)
 
 
 # Some of the following examples are taken from Flink SQL documentation
@@ -12,6 +13,7 @@ class TestSqlUtils(unittest.TestCase):
         self.assertFalse(is_ddl(simple_select))
         self.assertFalse(is_dml(simple_select))
         self.assertTrue(is_dql(simple_select))
+        self.assertFalse(is_metadata_query(simple_select))
 
     def test_select_hints(self):
         select = """select * from
@@ -23,6 +25,7 @@ class TestSqlUtils(unittest.TestCase):
         self.assertFalse(is_ddl(select))
         self.assertFalse(is_dml(select))
         self.assertTrue(is_dql(select))
+        self.assertFalse(is_metadata_query(select))
 
     def test_with_query(self):
         with_query = "WITH orders_with_total AS (\n    " \
@@ -32,18 +35,21 @@ class TestSqlUtils(unittest.TestCase):
         self.assertFalse(is_ddl(with_query))
         self.assertFalse(is_dml(with_query))
         self.assertTrue(is_dql(with_query))
+        self.assertFalse(is_metadata_query(with_query))
 
     def test_select_distinct(self):
         select = "SELECT DISTINCT id FROM Orders"
         self.assertFalse(is_ddl(select))
         self.assertFalse(is_dml(select))
         self.assertTrue(is_dql(select))
+        self.assertFalse(is_metadata_query(select))
 
     def test_select_values(self):
         select = "SELECT order_id, price FROM (VALUES (1, 2.0), (2, 3.1))  AS t (order_id, price)"
         self.assertFalse(is_ddl(select))
         self.assertFalse(is_dml(select))
         self.assertTrue(is_dql(select))
+        self.assertFalse(is_metadata_query(select))
 
     def test_select_cumulate(self):
         select = """SELECT * FROM TABLE(
@@ -55,6 +61,7 @@ class TestSqlUtils(unittest.TestCase):
         self.assertFalse(is_ddl(select))
         self.assertFalse(is_dml(select))
         self.assertTrue(is_dql(select))
+        self.assertFalse(is_metadata_query(select))
 
     def test_select_window_aggregation(self):
         select = """SELECT
@@ -67,6 +74,7 @@ GROUP BY
         self.assertFalse(is_ddl(select))
         self.assertFalse(is_dml(select))
         self.assertTrue(is_dql(select))
+        self.assertFalse(is_metadata_query(select))
 
     def test_select_over_aggregation(self):
         select = """SELECT order_id, order_time, amount,
@@ -80,12 +88,14 @@ GROUP BY
         self.assertFalse(is_ddl(select))
         self.assertFalse(is_dml(select))
         self.assertTrue(is_dql(select))
+        self.assertFalse(is_metadata_query(select))
 
     def test_create_table(self):
         create = "CREATE TABLE Orders (`user` BIGINT, product STRING, amount INT) WITH ('connector'='raw');"
         self.assertTrue(is_ddl(create))
         self.assertFalse(is_dml(create))
         self.assertFalse(is_dql(create))
+        self.assertFalse(is_metadata_query(create))
 
         create_not_exist = "\n\nCREATE TABLE IF NOT EXISTS " \
                            "Catalog.Orders (`user` BIGINT, product STRING, " \
@@ -93,18 +103,21 @@ GROUP BY
         self.assertTrue(is_ddl(create_not_exist))
         self.assertFalse(is_dml(create_not_exist))
         self.assertFalse(is_dql(create_not_exist))
+        self.assertFalse(is_metadata_query(create_not_exist))
 
     def test_create_catalog(self):
         create = "CREATE CATALOG Catalog WITH (key1  =  val1)"
         self.assertTrue(is_ddl(create))
         self.assertFalse(is_dml(create))
         self.assertFalse(is_dql(create))
+        self.assertFalse(is_metadata_query(create))
 
     def test_create_database(self):
         create = "CREATE DATABASE Catalog.Database WITH ('key1'  =  'val1')"
         self.assertTrue(is_ddl(create))
         self.assertFalse(is_dml(create))
         self.assertFalse(is_dql(create))
+        self.assertFalse(is_metadata_query(create))
 
     def test_create_view(self):
         create = "CREATE TEMPORARY VIEW IF NOT EXISTS MyView " \
@@ -112,76 +125,89 @@ GROUP BY
         self.assertTrue(is_ddl(create))
         self.assertFalse(is_dml(create))
         self.assertFalse(is_dql(create))
+        self.assertFalse(is_metadata_query(create))
 
     def test_create_function(self):
         create = "CREATE FUNCTION MyFunction AS pyflink.table.tests.test_udf.add LANGUAGE PYTHON"
         self.assertTrue(is_ddl(create))
         self.assertFalse(is_dml(create))
         self.assertFalse(is_dql(create))
+        self.assertFalse(is_metadata_query(create))
 
     def test_drop_table(self):
         drop = "  DROP\n TABLE \n\rOrders"
         self.assertTrue(is_ddl(drop))
         self.assertFalse(is_dml(drop))
         self.assertFalse(is_dql(drop))
+        self.assertFalse(is_metadata_query(drop))
 
     def test_drop_catalog(self):
         drop = "  DROP\n CATALOG IF EXISTS \n\rCatalog"
         self.assertTrue(is_ddl(drop))
         self.assertFalse(is_dml(drop))
         self.assertFalse(is_dql(drop))
+        self.assertFalse(is_metadata_query(drop))
 
     def test_drop_database(self):
         drop = "\n\nDROP DATABASE Production\n\nCASCADE\n\n"
         self.assertTrue(is_ddl(drop))
         self.assertFalse(is_dml(drop))
         self.assertFalse(is_dql(drop))
+        self.assertFalse(is_metadata_query(drop))
 
     def test_drop_view(self):
         drop = "DROP VIEW Catalog.Database.View"
         self.assertTrue(is_ddl(drop))
         self.assertFalse(is_dml(drop))
         self.assertFalse(is_dql(drop))
+        self.assertFalse(is_metadata_query(drop))
 
     def test_drop_function(self):
         drop = "DROP FUNCTION Function"
         self.assertTrue(is_ddl(drop))
         self.assertFalse(is_dml(drop))
         self.assertFalse(is_dql(drop))
+        self.assertFalse(is_metadata_query(drop))
 
     def test_alter_table(self):
         alter = "ALTER TABLE Orders RENAME TO Chaoses"
         self.assertTrue(is_ddl(alter))
         self.assertFalse(is_dml(alter))
         self.assertFalse(is_dql(alter))
+        self.assertFalse(is_metadata_query(alter))
 
         alter_properties = "ALTER TABLE Orders SET (\n  'connector' = 'other_connector'\n);"
         self.assertTrue(is_ddl(alter_properties))
         self.assertFalse(is_dml(alter_properties))
         self.assertFalse(is_dql(alter_properties))
+        self.assertFalse(is_metadata_query(alter_properties))
 
     def test_alter_view(self):
         alter = "ALTER VIEW MyView RENAME TO YourView"
         self.assertTrue(is_ddl(alter))
         self.assertFalse(is_dml(alter))
         self.assertFalse(is_dql(alter))
+        self.assertFalse(is_metadata_query(alter))
 
         alter_expression = "ALTER VIEW MyView AS (VALUES (1, 2, 3))"
         self.assertTrue(is_ddl(alter_expression))
         self.assertFalse(is_dml(alter_expression))
         self.assertFalse(is_dql(alter_expression))
+        self.assertFalse(is_metadata_query(alter_expression))
 
     def test_alter_database(self):
         alter = "ALTER DATABASE Production SET ('key1' = 'val2', 'key2' = 'val1'\r\t)"
         self.assertTrue(is_ddl(alter))
         self.assertFalse(is_dml(alter))
         self.assertFalse(is_dql(alter))
+        self.assertFalse(is_metadata_query(alter))
 
     def test_alter_function(self):
         alter = "ALTER FUNCTION AddFunction AS pyflink.table.tests.test_udf.sub"
         self.assertTrue(is_ddl(alter))
         self.assertFalse(is_dml(alter))
         self.assertFalse(is_dql(alter))
+        self.assertFalse(is_metadata_query(alter))
 
     def test_insert_from_select(self):
         insert = "INSERT INTO country_page_view " \
@@ -190,6 +216,7 @@ GROUP BY
         self.assertFalse(is_ddl(insert))
         self.assertTrue(is_dml(insert))
         self.assertFalse(is_dql(insert))
+        self.assertFalse(is_metadata_query(insert))
 
         execute_insert = "EXECUTE  INSERT INTO country_page_view " \
                          "PARTITION (date='2019-8-30', country='China')\n  " \
@@ -197,6 +224,7 @@ GROUP BY
         self.assertFalse(is_ddl(execute_insert))
         self.assertTrue(is_dml(execute_insert))
         self.assertFalse(is_dql(execute_insert))
+        self.assertFalse(is_metadata_query(execute_insert))
 
         execute_insert_newline = "EXECUTE \n\n\r\t\n INSERT INTO country_page_view\n\t " \
                                  "PARTITION (date='2019-8-30', country='China')\n  " \
@@ -204,6 +232,7 @@ GROUP BY
         self.assertFalse(is_ddl(execute_insert_newline))
         self.assertTrue(is_dml(execute_insert_newline))
         self.assertFalse(is_dql(execute_insert_newline))
+        self.assertFalse(is_metadata_query(execute_insert_newline))
 
         insert_overwrite_partition = "INSERT OVERWRITE country_page_view " \
                                      "PARTITION (date='2019-8-30')  \n  " \
@@ -211,6 +240,7 @@ GROUP BY
         self.assertFalse(is_ddl(insert_overwrite_partition))
         self.assertTrue(is_dml(insert_overwrite_partition))
         self.assertFalse(is_dql(insert_overwrite_partition))
+        self.assertFalse(is_metadata_query(insert_overwrite_partition))
 
         insert_append = "INSERT INTO country_page_view " \
                         "PARTITION (date='2019-8-30', country='China') (user)  " \
@@ -218,23 +248,27 @@ GROUP BY
         self.assertFalse(is_ddl(insert_append))
         self.assertTrue(is_dml(insert_append))
         self.assertFalse(is_dql(insert_append))
+        self.assertFalse(is_metadata_query(insert_append))
 
     def test_insert_values(self):
         insert = "INSERT INTO students\n  VALUES ('fred flintstone', 35, 1.28), ('barney rubble', 32, 2.32)"
         self.assertFalse(is_ddl(insert))
         self.assertTrue(is_dml(insert))
         self.assertFalse(is_dql(insert))
+        self.assertFalse(is_metadata_query(insert))
 
     def test_describe(self):
         describe = "DESCRIBE Orders"
         self.assertFalse(is_ddl(describe))
         self.assertFalse(is_dml(describe))
         self.assertTrue(is_dql(describe))
+        self.assertTrue(is_metadata_query(describe))
 
         desc = "DESC Database.Orders;"
         self.assertFalse(is_ddl(desc))
         self.assertFalse(is_dml(desc))
         self.assertTrue(is_dql(desc))
+        self.assertTrue(is_metadata_query(desc))
 
     def test_explain(self):
         explain = """
@@ -246,6 +280,7 @@ GROUP BY
         self.assertFalse(is_ddl(explain))
         self.assertFalse(is_dml(explain))
         self.assertTrue(is_dql(explain))
+        self.assertTrue(is_metadata_query(explain))
 
     def test_use(self):
         uses = [
@@ -258,6 +293,7 @@ GROUP BY
                 self.assertTrue(is_ddl(use))
                 self.assertFalse(is_dml(use))
                 self.assertFalse(is_dql(use))
+            self.assertFalse(is_metadata_query(use))
 
     def test_show_catalogs(self):
         shows = [
@@ -285,24 +321,28 @@ GROUP BY
                 self.assertFalse(is_ddl(show))
                 self.assertFalse(is_dml(show))
                 self.assertTrue(is_dql(show))
+                self.assertTrue(is_metadata_query(show))
 
     def test_load(self):
         load = "LOAD MODULE hive WITH ('hive-version' = '3.1.2')\n"
         self.assertTrue(is_ddl(load))
         self.assertFalse(is_dml(load))
         self.assertFalse(is_dql(load))
+        self.assertFalse(is_metadata_query(load))
 
     def test_unload(self):
         unload = "UNLOAD MODULE core"
         self.assertTrue(is_ddl(unload))
         self.assertFalse(is_dml(unload))
         self.assertFalse(is_dql(unload))
+        self.assertFalse(is_metadata_query(unload))
 
     def test_non_match(self):
         non_match = "  UNKNOWNABLEKEYWORD * FROM (SELECT * FROM SomeTable)"
         self.assertFalse(is_ddl(non_match))
         self.assertFalse(is_dml(non_match))
         self.assertFalse(is_dql(non_match))
+        self.assertFalse(is_metadata_query(non_match))
 
     def test_multiline_cell(self):
         multiline_str = "CREATE TABLE multiline (\n"\