+++ streaming_jupyter_integrations/magics.py
@@ -7,7 +7,7 @@ import signal
 import subprocess
 import sys
 from functools import wraps
-from typing import Any, Callable, Dict, Iterable, Tuple, Union, cast
+from typing import Any, Callable, Dict, Iterable, Optional, Tuple, Union, cast
 
 import nest_asyncio
 import pandas as pd
@@ -36,7 +36,8 @@ from .display import pyflink_result_kind_to_string
 from .jar_handler import JarHandler
 from .reflection import get_method_names_for
 from .sql_syntax_highlighting import SQLSyntaxHighlighting
-from .sql_utils import inline_sql_in_cell, is_dml, is_dql, is_query
+from .sql_utils import (inline_sql_in_cell, is_dml, is_dql, is_metadata_query,
+                        is_query)
 from .variable_substitution import CellContentFormatter
 from .yarn import find_session_jm_address
 
@@ -341,10 +342,18 @@ class Integrations(Magics):
         else:
             execution_result = self.st_env.execute_sql(stmt)
         print("Job started")
-        await self.__pull_results(execution_result, display_row_kind, is_dql(stmt))
+        # Pandas lib truncates view if the number of results exceeds the limit. The same applies to column width.
+        # If the query shows metadata, e.g. list of tables or list of columns, then no limit is applied.
+        pd_display_options = {
+            "display.max_rows": None if is_metadata_query(stmt) else 100,
+            "display.max_colwidth": None if is_metadata_query(stmt) else 100,
+        }
+        await self.__pull_results(execution_result, display_row_kind, is_dql(stmt), pd_display_options)
 
     async def __pull_results(self, execution_result: TableResult, display_row_kind: bool,
-                             display_results: bool) -> None:
+                             display_results: bool, pd_display_options: Optional[Dict[str, Any]] = None) -> None:
+        if not pd_display_options:
+            pd_display_options = {}
         # active polling
         while not self.interrupted:
             try:
@@ -355,7 +364,7 @@ class Integrations(Magics):
                     # if a select query has been executing then `wait` returns as soon as the first
                     # row is available. To display the results
                     print("Pulling query results...")
-                    await self.display_execution_result(execution_result, display_row_kind)
+                    await self.display_execution_result(execution_result, display_row_kind, pd_display_options)
                     return
                 else:
                     # if finished then return early even if the user interrupts after this
@@ -384,7 +393,8 @@ class Integrations(Magics):
         # usual happy path
         print("Execution successful")
 
-    async def display_execution_result(self, execution_result: TableResult, display_row_kind: bool) -> pd.DataFrame:
+    async def display_execution_result(self, execution_result: TableResult, display_row_kind: bool,
+                                       pd_display_options: Dict[str, Any]) -> pd.DataFrame:
         """
         Displays the execution result and returns a dataframe containing all the results.
         Display is done in a stream-like fashion displaying the results as they come.
@@ -393,6 +403,8 @@ class Integrations(Magics):
         columns = execution_result.get_table_schema().get_field_names()
         if display_row_kind:
             columns = ["row_kind"] + columns
+        for key, value in pd_display_options.items():
+            pd.set_option(key, value)
         df = pd.DataFrame(columns=columns)
         result_kind = execution_result.get_result_kind()