+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix execution of SHOW, EXPLAIN and DESCRIBE commands.
+
 ## [0.12.0] - 2022-11-30
 
 ### Added